requirements.txt
appnope==0.1.0
attrs==17.4.0
beautifulsoup4==4.6.0
bleach==2.1.2
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
DateTime==4.2
decorator==4.1.2
entrypoints==0.2.3
facebook-sdk==2.0.0
html5lib==1.0.1
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.0
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
matplotlib==2.1.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.13.3
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.2
pycurl==7.43.0.1
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.3.1
python-dateutil==2.6.1
pytz==2017.3
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
scipy==1.0.0
seaborn==0.8.1
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.0
zope.interface==4.4.3
